service/pyproject.toml
[tool.poetry]
name = "urban_heat"
version = "2024"
description = "A simple backend supporting the Urban Heat Map"
authors = ["Damien Allen <dev@dallen.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"

fastapi = "^0.110.1"
beanie = "^1.25.0"
pydantic = "^2.7.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }

aiofiles = "^23.2.1"
httpx = "^0.27.0"
rasterio = "^1.3.9"
geopandas = "^0.14.3"
pyyaml = "^6.0.1"
tqdm = "^4.66.2"
typer = "^0.12.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pre-commit = "^3.7"
ruff = "^0.4.3"
types-aiofiles = "^23.2.0.20240403"
types-pyyaml = "^6.0.12.20240311"


[tool.black]
line-length = 99

[tool.ruff]
line-length = 99

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]

[tool.mypy]
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
